backend/requirements.txt
# 核心依赖
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5

# 数据库
SQLAlchemy==1.4.23
# pymysql==1.0.2  # 注释掉 pymysql
mysqlclient==2.1.1  # 使用 mysqlclient 替代

# 认证
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
email-validator>=1.1.3
cryptography==39.0.1

# 环境变量
python-dotenv>=0.19.0

# HTTP 客户端
httpx==0.23.0
urllib3<2.0.0

# 工具
wheel>=0.37.0
setuptools>=57.0.0
numpy>=1.12.0

# 图像处理
Pillow>=8.0.0
matplotlib>=3.9.4

# 文件处理
aiohttp>=3.8.1
aiofiles>=0.8.0

# 移除不需要的 tensorboard 

# AI 服务
zhipuai>=1.0.7  # 添加智谱 AI SDK 